requirements.txt
requests>=2.25.1
pytest>=6.2.3
pytest-cov>=2.11.1
python-dotenv>=0.17.1